feat: add DOM helpers module
